git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rserpooldemo@1187 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ src/canvasrsernode.h
@@ -49,7 +49,7 @@ class CCanvasRSerNode : public QCanvasSprite
    QPopupMenu*       m_ContextMenu;
 
    private:
-   int               m_ZPosition;
+   double            m_ZPosition;
    CRSerPoolCanvas*  m_Canvas;
    CRSerPoolNode*    m_RSerNode;
